client: Introduce csimanager and csi fingerprints
This commit adds an initial implementation of CSI plugin fingerprinting,
via the introduction of the CSI manager, CSI Node Update events, and two
new fields on a Nomad Node.

It's structured after the other pluginmanager implementations, but
currently only fingerprints the NodeID from the plugin, and assumes that
all plugins are monolithic.

We need to expand this somewhat to seperate the registrations of
Controller plugins and Node plugins, but for initial prototyping this
should be ok.